CON-622: Extract `*ObservedBehavior` to separate files.
