+++ storage/src/main/scala/io/casperlabs/storage/dag/DagStorage.scala
@@ -21,12 +21,6 @@ import scala.util.Failure
 import scala.util.control.NoStackTrace
 import cats.Functor
 import com.github.ghik.silencer.silent
-import io.casperlabs.storage.dag.DagRepresentation.ObservedValidatorBehavior.{
-  Empty,
-  Equivocated,
-  Honest
-}
-import shapeless.tag.@@
 
 trait DagStorage[F[_]] {
 
@@ -283,119 +277,4 @@ object DagRepresentation {
   }
 
   def apply[F[_]](implicit ev: DagRepresentation[F]): DagRepresentation[F] = ev
-
-  final class EraObservedBehavior[A] private (
-      val data: Map[ByteString, Map[Validator, ObservedValidatorBehavior[A]]]
-  ) {
-    // Returns set of equivocating validators that are visible in the j-past-cone
-    // of the era.
-    def equivocatorsVisibleInEras(
-        keyBlockHashes: Set[ByteString]
-    ): Set[Validator] =
-      data
-        .collect {
-          case (keyBlockHash, lms) if keyBlockHashes(keyBlockHash) =>
-            lms.filter(_._2.isEquivocated).keySet
-        }
-        .flatten
-        .toSet
-
-    def keyBlockHashes: Set[ByteString] = data.keySet
-
-    def validatorsInEra(keyBlockHash: ByteString): Set[Validator] =
-      data.get(keyBlockHash).fold(Set.empty[Validator])(_.keySet)
-
-//    def latestMessagesInEra[F[_]: MonadThrowable: DagLookup](
-//        keyBlockHash: ByteString
-//    )(implicit ev: A =:= ByteString): F[Map[Validator, Set[Message]]] =
-//      data.get(keyBlockHash) match {
-//        case None => Map.empty[Validator, Set[Message]].pure[F]
-//        case Some(lms) =>
-//          lms.toList
-//            .traverse {
-//              case (v, observedBehavior) =>
-//                val lm: F[Set[Message]] = observedBehavior match {
-//                  case Empty => Set.empty[Message].pure[F]
-//                  case Honest(hash) =>
-//                    DagLookup[F].lookupUnsafe(hash.asInstanceOf[ByteString]).map(Set(_))
-//                  case Equivocated(m1, m2) =>
-//                    MonadThrowable[F].map2(
-//                      DagLookup[F].lookupUnsafe(m1.asInstanceOf[ByteString]),
-//                      DagLookup[F].lookupUnsafe(m2.asInstanceOf[ByteString])
-//                    ) { case (a, b) => Set(a, b) }
-//
-//                }
-//                lm.map(v -> _)
-//            }
-//            .map(_.toMap)
-//      }
-
-    def latestMessagesInEra(
-        keyBlockHash: ByteString
-    )(implicit ev: A =:= Message): Map[Validator, Set[Message]] =
-      data.get(keyBlockHash) match {
-        case None => Map.empty[Validator, Set[Message]]
-        case Some(lms) =>
-          lms.mapValues {
-            case Empty       => Set.empty[Message]
-            case Honest(msg) => Set(msg.asInstanceOf[Message])
-            case Equivocated(m1, m2) =>
-              Set(m1.asInstanceOf[Message], m2.asInstanceOf[Message])
-          }
-      }
-  }
-
-  sealed trait ObservedValidatorBehavior[+A] extends Product with Serializable {
-    def isEquivocated: Boolean
-  }
-
-  object ObservedValidatorBehavior {
-
-    case class Honest[A](msg: A) extends ObservedValidatorBehavior[A] {
-      override def isEquivocated: Boolean = false
-    }
-
-    case class Equivocated[A](m1: A, m2: A) extends ObservedValidatorBehavior[A] {
-      override def isEquivocated: Boolean = true
-    }
-
-    case object Empty extends ObservedValidatorBehavior[Nothing] {
-      override def isEquivocated: Boolean = false
-    }
-
-    implicit val functor: Functor[ObservedValidatorBehavior] =
-      new Functor[ObservedValidatorBehavior] {
-        def map[A, B](fa: ObservedValidatorBehavior[A])(f: A => B): ObservedValidatorBehavior[B] =
-          fa match {
-            case Honest(msg)         => Honest(f(msg))
-            case Equivocated(m1, m2) => Equivocated(f(m1), f(m2))
-            case Empty               => Empty
-          }
-      }
-  }
-
-  object EraObservedBehavior {
-    type LocalDagView[A] = EraObservedBehavior[A] @@ LocalView
-    type MessageJPast[A] = EraObservedBehavior[A] @@ MessageView
-
-    def local(data: Map[ByteString, Map[Validator, Set[Message]]]): LocalDagView[Message] =
-      apply(data).asInstanceOf[LocalDagView[Message]]
-
-    def messageJPast(data: Map[ByteString, Map[Validator, Set[Message]]]): MessageJPast[Message] =
-      apply(data).asInstanceOf[MessageJPast[Message]]
-
-    private def apply(
-        data: Map[ByteString, Map[Validator, Set[Message]]]
-    ): EraObservedBehavior[Message] =
-      new EraObservedBehavior(data.mapValues(_.map {
-        case (v, lms) =>
-          if (lms.isEmpty) v -> ObservedValidatorBehavior.Empty
-          else if (lms.size == 1) v -> ObservedValidatorBehavior.Honest(lms.head)
-          else v                    -> ObservedValidatorBehavior.Equivocated(lms.head, lms.drop(1).head)
-      }))
-
-    sealed trait LocalView
-
-    sealed trait MessageView
-  }
 }
